requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
cloudant==2.4.0
futures==3.0.5
kafka-python==1.3.3
packaging==16.8
pika==0.10.0
pkg-resources==0.0.0
pyparsing==2.2.0
redis==2.10.5
requests==2.13.0
requests-futures==0.9.4
six==1.10.0
